bcm-2xxx-rpi: mask build warning directly
Issue: LINCD-7632

Recently, Yocto require to use git protocol for the github download
address, otherwise report a warning. Raspberry firmware download
address starts with github, but it is released with the style of tar
ball, so there is a build warning reported when build it. If convert
to git protocol, it needs to create a mirror of firmware repo that is
more than 20GB, it will be a huge resource consumption. Therefore, mask
the warning directly.

Signed-off-by: Meng Li <Meng.Li@windriver.com>